Added function Get density by micro route id

+++ back-microroute_business-service/src/main/java/com/asb/back_micro_route_business_service/model/GgpMicroRoute.java
@@ -45,7 +45,7 @@ public class GgpMicroRoute {
     private Long  totalAforo;
 
     @Column(name = "density")
-    private Long  density;
+    private Double  density;
 
     @Column(name = "micro_route_sign")
     private String microRouteSign;
